Updating SP to long. Also fixed PetInfo packet format to avoid errors. Requires DP.

+++ L2J_Server/java/com/l2jserver/gameserver/model/actor/L2Character.java
@@ -4924,7 +4924,7 @@ public abstract class L2Character extends L2Object implements ISkillsHolder, IDe
 	 * @param addToExp
 	 * @param addToSp
 	 */
-	public void addExpAndSp(long addToExp, int addToSp)
+	public void addExpAndSp(long addToExp, long addToSp)
 	{
 		// Dummy method (overridden by players and pets)
 	}
